Move cluster name and region validator in validators layer
