+++ cli/src/pcluster/models/cluster.py
@@ -14,7 +14,6 @@
 #
 import json
 import logging
-import re
 import time
 from enum import Enum
 from typing import List
@@ -28,13 +27,7 @@ from common.aws.aws_resources import InstanceInfo, StackInfo
 from common.boto3.common import AWSClientError
 from pcluster.cli_commands.compute_fleet_status_manager import ComputeFleetStatus, ComputeFleetStatusManager
 from pcluster.config.config_patch import ConfigPatch
-from pcluster.constants import (
-    OS_MAPPING,
-    PCLUSTER_NAME_MAX_LENGTH,
-    PCLUSTER_NAME_REGEX,
-    PCLUSTER_STACK_PREFIX,
-    SUPPORTED_REGIONS,
-)
+from pcluster.constants import OS_MAPPING, PCLUSTER_STACK_PREFIX
 from pcluster.models.cluster_config import BaseClusterConfig, ClusterBucket, SlurmScheduling, Tag
 from pcluster.schemas.cluster_schema import ClusterSchema
 from pcluster.templates.cdk_builder import CDKTemplateBuilder
@@ -43,7 +36,6 @@ from pcluster.utils import (
     create_s3_bucket,
     generate_random_name_with_prefix,
     get_installed_version,
-    get_region,
     grouper,
     upload_resources_artifacts,
 )
@@ -337,10 +329,7 @@ class Cluster:
 
             # semantic validation
             if not suppress_validators:
-                validation_failures = self._validate_cluster_name()
-                validation_failures += self._validate_region()
-
-                validation_failures += config.validate()
+                validation_failures = config.validate()
                 for failure in validation_failures:
                     if failure.level.value >= FailureLevel(validation_failure_level).value:
                         # Raise the exception if there is a failure with a level greater than the specified one
@@ -353,31 +342,6 @@ class Cluster:
             raise ClusterActionError("Configuration is invalid", validation_failures=validation_failures)
 
         return config
-
-    def _validate_cluster_name(self):
-        failures = []
-        if not re.match(PCLUSTER_NAME_REGEX % (PCLUSTER_NAME_MAX_LENGTH - 1), self.name):
-            message = (
-                "Error: The cluster name can contain only alphanumeric characters (case-sensitive) and hyphens. "
-                "It must start with an alphabetic character and can't be longer "
-                f"than {PCLUSTER_NAME_MAX_LENGTH} characters."
-            )
-            LOGGER.error(message)
-            failures.append(ValidationResult(str(message), FailureLevel.ERROR))
-        return failures
-
-    @staticmethod
-    def _validate_region():
-        failures = []
-        region = get_region()
-        if region not in SUPPORTED_REGIONS:
-            failures.append(
-                ValidationResult(
-                    f"Region '{region}' is not yet officially supported by ParallelCluster",
-                    FailureLevel.ERROR,
-                )
-            )
-        return failures
 
     def _setup_cluster_bucket(self) -> ClusterBucket:
         """
