+++ cli/tests/pcluster/models/test_cluster.py
@@ -269,43 +269,3 @@ def _sort_tags(tags):
 
 def _sort_cfn_tags(tags):
     return sorted(tags, key=lambda tag: tag["Key"])
-
-
-@pytest.mark.parametrize(
-    "cluster_name, should_trigger_error",
-    [
-        ("ThisClusterNameShouldBeRightSize-ContainAHyphen-AndANumber12", False),
-        ("ThisClusterNameShouldBeJustOneCharacterTooLongAndShouldntBeOk", True),
-        ("2AClusterCanNotBeginByANumber", True),
-        ("ClusterCanNotContainUnderscores_LikeThis", True),
-        ("ClusterCanNotContainSpaces LikeThis", True),
-    ],
-)
-def test_validate_cluster_name(cluster_name, should_trigger_error, caplog):
-    error_msg = (
-        "Error: The cluster name can contain only alphanumeric characters (case-sensitive) and hyphens. "
-        f"It must start with an alphabetic character and can't be longer than {PCLUSTER_NAME_MAX_LENGTH} characters."
-    )
-    cluster = dummy_cluster(name=cluster_name)
-    failures = cluster._validate_cluster_name()
-    assert_that(failures).is_length(1 if should_trigger_error else 0)
-    if should_trigger_error:
-        assert_that(failures[0].message).is_equal_to(error_msg)
-
-
-@pytest.mark.parametrize(
-    "region, expected_message",
-    [
-        ("invalid-region", "Region 'invalid-region' is not yet officially supported "),
-        ("us-east-1", None),
-    ],
-)
-def test_region_validator(region, expected_message):
-    os.environ["AWS_DEFAULT_REGION"] = region
-
-    failures = dummy_cluster()._validate_region()
-    if expected_message:
-        assert_that(failures).is_length(1)
-        assert_that(failures[0].message).matches(expected_message)
-    else:
-        assert_that(failures).is_empty()
